requirements.txt
# ctransformers
# sentence-transformers
# pinecone
# langchain
# langchain-community
# langchain_pinecone
# flask
# pypdf
# python-dotenv


# -e .


Flask==2.3.3
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-pinecone==0.1.3
langchain-huggingface==0.0.3          # optional but recommended (replaces old HuggingFaceEmbeddings)
pinecone-client==5.0.1
sentence-transformers==2.2.2
CTransformers==0.2.27
torch==2.1.2                          # or 2.2.0+ if you have CUDA
accelerate==0.33.0
transformers==4.41.2
huggingface-hub==0.25.2               # CRITICAL: avoids the cached_download error
pypdf==4.2.0
tiktoken==0.7.0